[FEAT] #62 - 메인 화면 전환 기능 완료 및 간편 리뷰 기능 추가
